add some code

+++ Vocompfest/2018/Final/g.cpp
@@ -35,7 +35,7 @@ struct line{
 	}
 };
 
-int cross(point a, point b){
+ll cross(point a, point b){
 	return a.x*b.y - a.y*b.x;
 }
